+++ e2e/e2e-app/playwright.config.ts
@@ -4,15 +4,17 @@ import path from 'path';
 import { fileURLToPath } from 'url';
 import { PackageName } from './src/testConfig';
 
-export const targets: { packageName: PackageName; port: string }[] = [
-  // { packageName: 'e2e-alpine' },
-  { packageName: 'e2e-angular' },
-  { packageName: 'e2e-qwik' },
-  { packageName: 'e2e-react' },
-  { packageName: 'e2e-solid' },
-  { packageName: 'e2e-svelte' },
-  { packageName: 'e2e-vue3' },
-].map(({ packageName }, i) => {
+export const targets: { packageName: PackageName; port: number }[] = (
+  [
+    // { packageName: 'e2e-alpine' },
+    { packageName: 'e2e-angular' },
+    { packageName: 'e2e-qwik' },
+    { packageName: 'e2e-react' },
+    { packageName: 'e2e-solid' },
+    { packageName: 'e2e-svelte' },
+    { packageName: 'e2e-vue3' },
+  ] as const
+).map(({ packageName }, i) => {
   const port = 1234 + i;
 
   return {