+++ packages/core/src/__tests__/test-generator.ts
@@ -1,3 +1,4 @@
+import { describe, test } from 'vitest';
 import { MitosisComponent, createTypescriptProject, parseSvelte } from '..';
 import { parseJsx } from '../parsers/jsx';
 import { Target } from '../types/config';
@@ -119,6 +120,8 @@ const builderRenderContent = getRawFile('./data/blocks/builder-render-content.ra
 const rootFragmentMultiNode = getRawFile('./data/blocks/root-fragment-multi-node.raw.tsx');
 const renderContentExample = getRawFile('./data/render-content.raw.tsx');
 const onClickToPressable = getRawFile('./data/react-native/onclick-to-pressable.raw.tsx');
+const inputToTextInputRN = getRawFile('./data/react-native/text-input.raw.tsx');
+
 type Tests = { [index: string]: RawFile };
 
 const SVELTE_SYNTAX_TESTS: Tests = {
@@ -142,6 +145,7 @@ const SVELTE_SYNTAX_TESTS: Tests = {
 
 const REACT_NATIVE_TESTS: Tests = {
   onClickToPressable,
+  inputToTextInputRN,
 };
 
 const BASIC_TESTS: Tests = {