+++ packages/docs/vite.config.ts
@@ -47,7 +47,7 @@ export default defineConfig(({ command, mode }): UserConfig => {
     optimizeDeps: {
       // Put problematic deps that break bundling here, mostly those with binaries.
       // For example ['better-sqlite3'] if you use that in server functions.
-      exclude: [],
+      // exclude: [],
     },
     // This tells Vite how to bundle the server code.
     ssr: